+++ src/features/seller/components/index.ts
@@ -1,3 +1,7 @@
 export { SellerProductCard } from "./SellerProductCard";
 export { SellerOrdersView } from "./SellerOrdersView";
 export { SellerProductsView } from "./SellerProductsView";
+export { SellerDashboardView } from "./SellerDashboardView";
+export { SellerStatCard } from "./SellerStatCard";
+export { SellerQuickActions } from "./SellerQuickActions";
+export { SellerRecentListings } from "./SellerRecentListings";